feat: splitting REST API controllers tp different packages
Refs: #232
