+++ app/src/test/java/ch/sbb/polarion/extension/generic/service/PolarionServiceTest.java
@@ -4,8 +4,8 @@ import ch.sbb.polarion.extension.generic.exception.ObjectNotFoundException;
 import ch.sbb.polarion.extension.generic.fields.FieldType;
 import ch.sbb.polarion.extension.generic.fields.model.FieldMetadata;
 import ch.sbb.polarion.extension.generic.fields.model.Option;
-import ch.sbb.polarion.extension.generic.polarion.PlatformContextMockExtension;
-import ch.sbb.polarion.extension.generic.polarion.TransactionalExecutorExtension;
+import ch.sbb.polarion.extension.generic.test_extensions.PlatformContextMockExtension;
+import ch.sbb.polarion.extension.generic.test_extensions.TransactionalExecutorExtension;
 import ch.sbb.polarion.extension.generic.util.AssigneeUtils;
 import ch.sbb.polarion.extension.generic.util.TestUtils;
 import com.polarion.alm.projects.IProjectService;
