+++ app/src/test/java/ch/sbb/polarion/extension/generic/test_extensions/CustomExtensionMockInjector.java
@@ -1,4 +1,4 @@
-package ch.sbb.polarion.extension.generic.polarion;
+package ch.sbb.polarion.extension.generic.test_extensions;
 
 import org.junit.jupiter.api.extension.ExtensionContext;
